There is some weird shit going on after updating boards/libraries
For some reason, I need to manually call esp_wifi_start() after coming out of light sleep for WIFI to work. Found on the interwebs. Don't know why it doesn't work without it because it should so.......

Also removed the cert stuff because I think there are multiple certs right now with the API changes. Increased the interval to limit API access.