Adding time check before API call
Because we're limited to 1000 api calls a month, this code limits calls during daylight hours. Currently its hardcoded to PDT, I may solve this later to actually check the sunset and handle time zones but lets be real, there's not much solar output after 7pm and before 7am......let alone am I even awake that early.

So this is cobbled together to cut the API calls in half.